rolex: Tune VM memory parameters
dalvik.vm.heapminfree: 4m -> 512kb, reduce the minimum reclaim value,
this will prevent stalls in resource-constrained scenarios.

dalvik.vm.heapstartsize: 16m -> 8m, save a few mb on low-ram usage
services and apps, trading it off for the time it takes to increase the
heap in the cases where it would be necessary.

Change-Id: I6b3bacfc061f7d2ec6cce57738a79dc5861894be
Signed-off-by: adi153 <singhmanohar6699@gmail.com>